ranking FUNCIONAL

+++ Entidades/Personagens/jogador.hpp
@@ -26,7 +26,7 @@ namespace Entidades
             void colidir(Entidade* pE, int a);
 
             void set_tempo(double x) {tempo = x;}
-            double get_tempo() {return tempo;}
+            double get_tempo() {return (int)tempo;}
 
             void set_nome(std::string a)  {nome = a;}
             string get_nome(){return nome;}
